refactor: move signup and login controller to users folder and move import protocols to external file
