+++ src/main/factory/signup/signup-controller.ts
@@ -1,7 +1,7 @@
 import { DbAddAccount } from '../../../data/usecases/db/add-account/db-add-account'
 import { BcryptAdapter } from '../../../infra/usecases/cryptography/bcrypt/bcrypt-adapter'
 import { PgAccountRepository } from '../../../infra/usecases/db/postgres/account/pg-account-repository'
-import { SignUpController } from '../../../presentation/controller/signup/signup-controller'
+import { SignUpController } from '../../../presentation/controller/user/signup/signup-controller'
 import { makeSignUpValidation } from './signup-validation'
 
 export const makeSignUpController = (): SignUpController => {
