--- src/presentation/controller/login/login-controller-protocols.ts
@@ -1,3 +1,0 @@
-export * from '../../protocols/controller'
-export * from '../../protocols/http'
-export * from '../../helpers/http-helper'
