--- src/presentation/controller/signup/signup-controller-protocols.ts
@@ -1,4 +1,0 @@
-export * from '../../protocols/controller'
-export * from '../../protocols/http'
-export * from '../../helpers/http-helper'
-export * from '../../errors/email-already-exist-error'
