+++ src/presentation/controller/user/signup/signup-controller-protocols.ts
@@ -1,0 +1,5 @@
+export * from '../../../protocols/controller'
+export * from '../../../protocols/http'
+export * from '../../../helpers/http-helper'
+export * from '../../../errors/email-already-exist-error'
+export * from '../../../protocols/validation'
